refactor(snmptraps): migrate status component to v2 structure
Move from v1 layout (component.go + statusimpl/) to v2 layout
(def/ + impl/ + fx/) with Requires/Provides pattern.
