+++ comp/snmptraps/status/def/component.go
